I improved the API for the BaseMatchCallback class.
Specifically, I made the output() and getAsString() methods more
generic.
Now they cover all use cases.
I was also able to get rid of the overridden and overloaded methods
in the derived classes. Now only the base class methods are used.
I also changed the places where these methods are called to be
consistent and usable with the new API.
The behavior isn't changed, the code works as it did before,
just the code has been restructured to be more compact, idiomatic,
and maintainable.