Now I have modified the program to not use the class StaticAnalysisASTConsumer any more.
This change is inside StaticAnalysisAction::CreateASTConsumer().
There are two possible different implementations.
 - Either a MultiplexConsumer is returned.
 - Xor a StaticAnalysisASTConsumer is returned.
I have changed the current version to return a MultiplexConsumer.
I do not know if the StaticAnalysisASTConsumer will be needed later in
the future, maybe with some modifications.
So I just have saved it for the time being without deleting it.