I updated the API of the Base Callback Class.
I found some common functionality that was redundantly duplicated
in various derived classes, and I put that into the base class.
These were instances of when a SourceLocation as an offset from
another SourceLocation, based on seeking for a character, was being
generated.
The resulting abstraction is BaseMatchCallback::getCharOffsetLoc().
All Callback classes implementing similar functionality were trimmed
and they now call that method of the base class instead.