I created a new derived Match Callback class which removes the initialize function.
Based on the name of the currently processed source file, the
code between the { } of the initialize function is removed.

NOTE: This refactoring is applied successfully, but it generates
a Segmentation fault when the model is run.
It turns out that removing the initialization function completely
is not necessarily the best idea. It apparently does some important
tasks.