I created the module for removing Indirect Recursion in one of the models.
There are two CallBack classes here.
CreateMinorStepFunction creates the step_minor() function as an
initial duplicate of the step() function, and it replaces all calls
to step() with all calls to step_minor()

SeparateStepFunctions removes all if (MAJOR_TIME_STEP) {} statements
in the step_minor() function, and it removes all if (MINOR_TIME_STEP) {}
statements in the step() function.