Since adding Remove Recursion Module, had broken Remove Unused Variables Module, it had to be fixed.
The problem was that CreateMinorStepFunction
created another step function with the same variables.
So the unused variables would have been counted twice, and not removed.
I fixed this issue by keeping track of both the variable itself
and the function inside of which it resides.

This commit represents a fix into the Remove Unused Variables Module
to make it work better with Remove Recursion Module, and also
other modules in the future.
This version is more stable, sophisticated, and safer than the previous one.