I updated RemoveMemset module.
Now it deletes any calls to memset() that are in the initialize
function.
And it attempts to replace all calls to memset() that are not in
the initialize function with equivalent expressions.

This gets rid of the former issue when an entire object was
memset() into all zeoes, inside the memset() function.